requirements.txt
requests
pykrx
isodate
openai
playwright
python-dotenv
pytz
apscheduler
redis
tenacity
websockets